+++ backends-velox/src/main/scala/org/apache/gluten/backendsapi/velox/VeloxBackend.scala
@@ -41,6 +41,8 @@ import org.apache.spark.sql.hive.execution.HiveFileFormat
 import org.apache.spark.sql.internal.SQLConf
 import org.apache.spark.sql.types._
 
+import org.apache.hadoop.fs.Path
+
 import scala.util.control.Breaks.breakable
 
 class VeloxBackend extends Backend {
@@ -70,19 +72,19 @@ object VeloxBackendSettings extends BackendSettingsApi {
 
   val MAXIMUM_BATCH_SIZE: Int = 32768
 
-  override def supportFileFormatRead(
+  override def validateScan(
       format: ReadFileFormat,
       fields: Array[StructField],
       partTable: Boolean,
       rootPaths: Seq[String],
       paths: Seq[String]): ValidationResult = {
-    val filteredRootPaths = FileIndexUtil.distinctRootPaths(rootPaths)
+    val filteredRootPaths = distinctRootPaths(rootPaths)
     if (
       !filteredRootPaths.isEmpty && !VeloxFileSystemValidationJniWrapper
         .allSupportedByRegisteredFileSystems(filteredRootPaths.toArray)
     ) {
       return ValidationResult.failed(
-        s"Schema of [$filteredRootPaths] is not supported by registered file systems.")
+        s"Scheme of [$filteredRootPaths] is not supported by registered file systems.")
     }
     // Validate if all types are supported.
     def validateTypes(validatorFunc: PartialFunction[StructField, String]): ValidationResult = {
@@ -186,6 +188,17 @@ object VeloxBackendSettings extends BackendSettingsApi {
           .getRawTypeString(metadata)
           .getOrElse(stringType.catalogString))
       .isDefined
+  }
+
+  def distinctRootPaths(paths: Seq[String]): Seq[String] = {
+    // Skip native validation for local path, as local file system is always registered.
+    // For evey file scheme, only one path is kept.
+    paths
+      .map(p => (new Path(p).toUri.getScheme, p))
+      .groupBy(_._1)
+      .filter(_._1 != "file")
+      .map(_._2.head._2)
+      .toSeq
   }
 
   override def supportWriteFilesExec(
