+++ gluten-core/src/main/scala/org/apache/gluten/utils/FileIndexUtil.scala
@@ -18,24 +18,11 @@ package org.apache.gluten.utils
 
 import org.apache.spark.sql.execution.datasources._
 
-import java.net.URI
-
 object FileIndexUtil {
   def getRootPath(index: FileIndex): Seq[String] = {
     index.rootPaths
       .filter(_.isAbsolute)
       .map(_.toString)
-      .toSeq
-  }
-
-  def distinctRootPaths(paths: Seq[String]): Seq[String] = {
-    // Skip native validation for local path, as local file system is always registered.
-    // For evey file schema, only one path is kept.
-    paths
-      .map(p => (URI.create(p).getScheme, p))
-      .groupBy(_._1)
-      .filter(_._1 != "file")
-      .map(_._2.head._2)
       .toSeq
   }
 }
